+++ Meitner/System/SubSystems/Simulation/SimSystem/src/NativeMemory.h
@@ -31,7 +31,7 @@
 // OF SUCH DAMAGE.
 // 
 // Title : Simulation sub-system
-// Rev.  : 1/30/2023 Mon (clonextop@gmail.com)
+// Rev.  : 2/1/2023 Wed (clonextop@gmail.com)
 //================================================================================
 #ifndef __NATIVE_MEMORY_H__
 #define __NATIVE_MEMORY_H__
@@ -40,11 +40,11 @@
 class NativeSystemMemory : public IMemoryNative {
 	BYTE*	m_pMem;
 public:
-	NativeSystemMemory(DWORD dwByteSize);
+	NativeSystemMemory(UINT64 dwByteSize);
 	virtual ~NativeSystemMemory(void);
 
 	virtual void Release(void);
 	virtual BYTE* Virtual(void);
-	virtual bool Flush(DWORD dwOffset, DWORD dwPhyAddress, DWORD dwByteSize, bool bWrite);
+	virtual bool Flush(UINT64 dwOffset, UINT64 dwPhyAddress, UINT64 dwByteSize, bool bWrite);
 };
 #endif//__NATIVE_MEMORY_H__
